patch test-matrix_to_df to deal with systems without long.double
Prior to this change the test using `identical` would fail on a system
with no long double support.

This change replaces the `identical` call with `all.equal` and that
should all the testing set to pass on systems without long.double.

Also, a similar issue with the M1 Mac was reported and we hope this
patch address this issue too.